+++ compiler/src/instructions/JumpInstruction.ts
@@ -1,4 +1,6 @@
 import { IBindableValue, IValue } from "../types";
+import { LiteralValue } from "../values";
+import { JumpOutValue } from "../values/JumpOutValue";
 import { InstructionBase } from "./InstructionBase";
 
 export enum EJumpKind {
@@ -20,5 +22,21 @@ export class JumpInstruction extends InstructionBase {
     right: IValue | null = null
   ) {
     super("jump", address, kind, left, right);
+  }
+
+  /**
+   * Works with the `JumpOutValue` class to determine
+   * wether a jump should be added if the expression wasn't jump compressed.
+   */
+  static or(test: IValue, out: JumpOutValue) {
+    if (test === out) return [];
+    return [
+      new JumpInstruction(
+        out.address,
+        EJumpKind.Equal,
+        test,
+        new LiteralValue(+out.whenTrue)
+      ),
+    ];
   }
 }